Removed unnecessary code Added HeAACv1, HeAACv2 as separate files Adjusted imports
