+++ bitmovin/services/codecconfigurations/aac_codec_configuration_service.py
@@ -1,4 +1,4 @@
-from bitmovin.resources.models import AACCodecConfiguration, HeAACv1CodecConfiguration, HeAACv2CodecConfiguration
+from bitmovin.resources.models import AACCodecConfiguration, HeAACv2CodecConfiguration
 from ..rest_service import RestService
 
 
@@ -7,17 +7,3 @@ class AAC(RestService):
 
     def __init__(self, http_client):
         super().__init__(http_client=http_client, relative_url=self.BASE_ENDPOINT_URL, class_=AACCodecConfiguration)
-
-
-class HeAACv1(RestService):
-    BASE_ENDPOINT_URL = 'encoding/configurations/audio/he-aac-v1'
-
-    def __init__(self, http_client):
-        super().__init__(http_client=http_client, relative_url=self.BASE_ENDPOINT_URL, class_=HeAACv1CodecConfiguration)
-
-
-class HeAACv2(RestService):
-    BASE_ENDPOINT_URL = 'encoding/configurations/audio/he-aac-v2'
-
-    def __init__(self, http_client):
-        super().__init__(http_client=http_client, relative_url=self.BASE_ENDPOINT_URL, class_=HeAACv2CodecConfiguration)
